+++ src/config.lua
@@ -12,20 +12,23 @@ local config = {
     DisplayWidth = 400,
     BackgroundColor = { 0.090, 0.055, 0.157, 0.6 },
     ShowIcons = true,
-    SplitOmega = true
+    SplitOmega = true,
+    OmegaIndicator = "{!Icons.Omega_NoTooltip}"
 }
 
 local configDesc = {
-    PollingInterval = "How often to update bars, in seconds",
-    TrainingRoomClearTime = "How soon to clear the window in Skelly's room, in seconds",
-    InitialY = "Y position of meter",
-    XPosition = "X position of meter",
-    YPositionIncrement = "Vertical spacing for each bar, negative number",
-    Margin = "Width around the thing",
-    DisplayWidth = "Width of the meter background",
-    BackgroundColor = "Color of the meter background",
-    ShowIcons = "Show God icons next to ability names",
-    SplitOmega = "Show Omega attacks/specials in separate bars"
+    PollingInterval = "How often to update bars, in seconds.",
+    TrainingRoomClearTime = "How soon to clear the window in Skelly's room, in seconds.",
+    InitialY = "Y position of meter.",
+    XPosition = "X position of meter.",
+    YPositionIncrement = "Vertical spacing for each bar, negative number.",
+    Margin = "Used for meter background scaling.",
+    DisplayWidth = "Width of the meter background.",
+    BackgroundColor = "Color of the meter background.",
+    ShowIcons = "Show God icons next to ability names.",
+    SplitOmega = "Show Omega attacks/specials in separate bars.",
+    OmegaIndicator =
+    "If SplitOmega is enabled, this will prefix the given attack/special name. Can use plain text if the default icon causes flickering."
 }
 
 return config, configDesc